Feat: 핸들러 메서드 및 dto 작성
- GET /api/notifications와 매핑되는 핸들러 메서드를 작성했습니다.
- 위 api와 관련된 request, response dto를 작성했습니다.

+++ src/main/java/com/codeit/mission/deokhugam/notification/dto/CursorPageResponseNotificationDto.java
@@ -1,6 +1,5 @@
 package com.codeit.mission.deokhugam.notification.dto;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Builder;
@@ -11,7 +10,7 @@ public record CursorPageResponseNotificationDto(
     String nextCursor,
     LocalDateTime nextAfter,
     int size,
-    BigInteger totalElements,
+    long totalElements,
     boolean hasNext
 ) {
